* Moving all progressbars to a single file
